Upgrade to hibernate-validator 6.0.7.Final
There were multiple versions of this dependency on the classpath,
which were causing odd issues with the @NotEmpty annotation. In some
cases the build was reporting it was deprecated, but then when using
the JSR303 annotation, the build would fail. Standardizing on the
Spring Boot managed version and updating all usage to use JSR303
annotations results in a happy build and no deprecation warnings.