Add model object for ServiceMetadata and improve imageUrl support
This commit adds a new `ServiceMetadata` model object that allows for the
service offering metadata to be more easily configured in Spring
configuration properties. While these properties are not specifically part of
the OSB API spec, they are recommended for consistency.

Additionally, images located on the class path can now be configured in the
`metadata.imageUrlResource` property. The image will be Base64 encoded and the
value provided in the standard `metadata.imageUrl` for the service offering.

Resolves #207