Return HTTP 400 Bad Request for unknown instance in last operation
v2.15 of the OSB API spec is a little unclear about how to handle the
scenario where a request is made for the last operation of an instance id
that does not exist and never existed. It does specify a 400 Bad Request
for "malformed or missing mandatory data". In this case an unknown instance
id is provided, so it's the best option for response code.

The spec is improved in v2.16. There, it adds a 404 Not Found for this
scenario. We'll change this when we move to the newer spec.

Resolves #306