Add fields to indicate if instance is usable and operation is repeatable on failures
Add instance_usable and update_repeatable boolean fields to
GetLastServiceOperationResponse and ErrorMessage in order to capture the
following information:

instance_usable - boolean - If an update or deprovisioning operation failed,
this flag indicates whether or not the Service Instance is still usable. If
true, the Service Instance can still be used, false otherwise. This field
MUST NOT be present for errors of other operations. Defaults to true.

update_repeatable - boolean - If an update operation failed, this flag
indicates whether this update can be repeated or not. If true, the same
update operation MAY be repeated and MAY succeed; if false, repeating the
same update operation will fail again. This field MUST NOT be present for
errors of other operations. Defaults to true.

Resolves #353