Support 200 and 409 response codes for existing binding requests
This commit deprecates the `isBindingExisted` property and adds a new
`BindingStatus` enum to support the three possible responses when creating
a service instance binding. A 200 OK will be returned if a binding exists
with identical parameters, and a 409 CONFLICT will be returned if a binding
exists with different parameters. It is the responsibility of the
implementing application to determine which of these is applicable.

Resolves #321